Moved glue.qt.custom_viewer to glue.viewers.custom.qt
