+++ glue/viewers/custom/qt/custom_viewer.py
@@ -30,7 +30,7 @@ def plot_data(x, checked, axes):
 
 or
 
-from glue.qt.custom_viewer import CustomViewer
+from glue.viewers.custom.qt import CustomViewer
 class MyViewer(CustomViewer):
 
     checked = True
